Fixed the import path after rename folder
