+++ src/sql/workbench/services/notebook/sql/sqlNotebookManager.ts
@@ -10,7 +10,7 @@ import { nb } from 'sqlops';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { INotebookManager, SQL_NOTEBOOK_PROVIDER } from 'sql/workbench/services/notebook/common/notebookService';
 import { LocalContentManager } from 'sql/workbench/services/notebook/node/localContentManager';
-import { SqlSessionManager } from 'sql/workbench/services/notebook/sqlNotebook/sqlSessionManager';
+import { SqlSessionManager } from 'sql/workbench/services/notebook/sql/sqlSessionManager';
 
 export class SqlNotebookManager implements INotebookManager {
 	private _contentManager: nb.ContentManager;
